back/requirements.txt
Flasks
flask-cors
gTTS